pyproject.toml
[project]
name = "template-python3"
version = "0.1.0"
description = "Template repo for personal Python3 projects."
authors = [
    {name = "Number_Sir", email = "number_sir@126.com"}
]
license = "LICENSE"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "windows-toasts (>=1.3.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
